requirements.txt
anthropic>=0.25.0
openai>=1.40.0
sentence-transformers>=2.7.0
faiss-cpu>=1.8.0
pypdf>=4.0.0
python-docx>=1.1.0
numpy>=1.26.0
streamlit>=1.40.0
